Fixed test request signature creation
