+++ tests/conftest.py
@@ -1,6 +1,17 @@
-import pytest
 import json
+import hashlib
+import hmac
+import pytest
 from slackeventsapi import SlackEventAdapter
+
+
+def create_signature(secret, timestamp, data):
+    req = str.encode('v0:' + str(timestamp) + ':') + str.encode(data)
+    request_signature= 'v0='+hmac.new(
+        str.encode(secret),
+        req, hashlib.sha256
+    ).hexdigest()
+    return request_signature
 
 
 def load_event_fixture(event, as_string=True):
@@ -23,7 +34,8 @@ def pytest_namespace():
     return {
         'reaction_event_fixture': load_event_fixture('reaction_added'),
         'url_challenge_fixture': load_event_fixture('url_challenge'),
-        'bad_token_fixture': event_with_bad_token()
+        'bad_token_fixture': event_with_bad_token(),
+        'create_signature': create_signature
     }
